refactor : component 구조
